Changing EnvelopeDetails to Envelope
